refactor(indexer): decompose index_directory into phase functions
Split the ~540-line index_directory orchestrator into named phase fns,
leaving it a thin driver over the incremental-index pipeline:

- walk_candidates (Phase 1) lives in walk.rs alongside the WalkFilter it
  applies; parse_candidates / store_parsed_file / resolve_and_finalize
  (Phases 2-4) live in a new pass.rs module.
- The single-transaction atomicity invariant is preserved: the tx guard
  stays owned by index_directory and never crosses a fn boundary; the
  phase helpers take &Database and call only the *_in_tx batch helpers,
  so every write joins the one transaction.

Behavior-preserving move (1581 workspace tests unchanged). Phase fns
share the &mut IndexResult convention and carry #[must_use].
